Modify `query` on-the-fly
Add a `assignment_expression` query expression on-the-fly.  This is
e.g. for JavaScript where standard assignments do *not* count as
`definition.var` but only `var foo=1;` expressions.

We need to probe this as the queries are checked by `tree_sitter` and
not resilient.  T.i. if the grammar does not define a
`assignment_expression`, the query is not allowed to use it either.
(Why is that so strict?  Is that pluggable?)